+++ configs/tf-1d/damping.yaml
@@ -14,6 +14,12 @@ models:
       out_size: 1
       width_size: 8
 
+units:
+  laser wavelength: 351nm
+  electron temperature: 2000eV
+  density for collisions: 1.5e21/cc
+  Z: 10
+  Zp: 10
 
 grid:
     nx: 16
