+++ configs/tf-1d/epw.yaml
@@ -55,6 +55,7 @@ physics:
       kld: 0.3
       nuee: 1.0e-09
   landau_damping: true
+
 save:
   t:
     nt: 1000
@@ -64,3 +65,10 @@ save:
     nx: 16
     xmax: 19.634954084936208
     xmin: 0.0
+
+units:
+  laser wavelength: 351nm
+  electron temperature: 2000eV
+  density for collisions: 1.5e21/cc
+  Z: 10
+  Zp: 10
