+++ configs/tf-1d/wp.yaml
@@ -73,3 +73,10 @@ save:
     nx: 256
     xmax: 1000
     xmin: 0.0
+
+units:
+  laser wavelength: 351nm
+  electron temperature: 2000eV
+  density for collisions: 1.5e21/cc
+  Z: 10
+  Zp: 10
