+++ tests/test_tf1d/configs/resonance.yaml
@@ -13,6 +13,13 @@ grid:
 
 models: False
 
+units:
+  laser wavelength: 351nm
+  electron temperature: 2000eV
+  density for collisions: 1.5e21/cc
+  Z: 10
+  Zp: 10
+
 save:
     t:
       tmin: 0.0
