+++ tests/test_tf1d/configs/resonance_search.yaml
@@ -6,6 +6,13 @@ mlflow:
 
 adjoint: Backsolve
 
+units:
+  laser wavelength: 351nm
+  electron temperature: 2000eV
+  density for collisions: 1.5e21/cc
+  Z: 10
+  Zp: 10
+
 grid:
     nx: 16
     xmin: 0.0
